Add OKCAR option to USB device mode
When the device switches to this option, the USB device descriptor is consistent with the iPhone, allowing the car system to initiate a CarPlay connection.

These modifications only provide switching options at the upper level, while the main USB descriptor simulation behavior occurs in the kernel. The kernel needs to combine multiple USB gadget functions to achieve a USB descriptor similar to that of the iPhone. The init.okcar.rc file responds to 'on property:sys.usb.config=okcar', which notifies the kernel to start the simulation.

Change-Id: Icec615762582c76a8cfd6e16c057d44d51ef21b5